Update on adapting with categories

+++ smogn/over_sampling.py
@@ -130,8 +130,9 @@ def over_sampling(
             if data.dtypes[j] == 'category':
                 uniques = data.iloc[:, j].cat.categories
                 codes= data.iloc[:, j].cat.codes.values
+                # clean the data.iloc[:, j] such that it is not a category and be int
                 data.iloc[:, j] = None
-                data.iloc[:, j] = data.iloc[:, j].astype('int64')
+                data.iloc[:, j] = data.iloc[:, j].astype('float')
                 data.iloc[:, j] = pd.Categorical(codes)
                 #data.iloc[:, j] = pd.Categorical(data.iloc[:, j].cat.codes)
             else:
